+++ packages/activity-manager/src/handlers.ts
@@ -1,6 +1,6 @@
-import { Logger } from "./logger";
-import { ActivityRunnable } from "./types";
-import { ActivityEvents } from "@sugarch/bc-event-handler";
+import { Logger } from './logger';
+import { ActivityRunnable } from './types';
+import { ActivityEvents } from '@sugarch/bc-event-handler';
 
 const deRepeats = new Set<string>();
 
@@ -9,13 +9,13 @@ const deRepeats = new Set<string>();
  * @param name - Name of the activity
  * @param handler - Handler to execute when the activity occurs
  */
-export function pushHandler(name: string, handler: ActivityRunnable): void {
+export function pushHandler (name: string, handler: ActivityRunnable): void {
     if (deRepeats.has(name)) {
         Logger.warn(`Handler for ${name} already exists, skipping`);
         return;
     }
 
-    ActivityEvents.on(handler.mode ?? 'AnyOnSelf', name as ActivityName, (sender, player, info) => {
+    ActivityEvents.on(handler.mode ?? 'SelfInvolved', name as ActivityName, (sender, player, info) => {
         handler.run?.(player, sender, info);
     });
 }