+++ packages/activity-manager/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sugarch/bc-activity-manager
 
+## 0.2.0
+
+### Minor Changes
+
+-   Refactor activity trigger, use 'SelfInvolved' as default trigger mode.
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @sugarch/bc-mod-types@0.6.0
+    -   @sugarch/bc-mod-hook-manager@0.2.11
+
 ## 0.1.6
 
 ### Patch Changes